Added .csv files containing generated mock data
This data does not include all columns. games needs an additional column of avg_rating that contains the average of all ratings referencing that game's id. orders needs total_price (a sum of all price_at_purchase from order details referencing that order's id) and transaction_id (the id of the order that references that transaction)